Make sure default ALT0 hardware PWM pin mode isn't overwritten
RPi.GPIO only supports `IN` and `OUT` pin modes but not the special PWM mode `ALT0`. In case hardware PWM is properly configured, the system will boot up with the pin in the correct mode. However, overwriting the pin mode will get rid of that configuration and it will render the pin useless for hardware PWM.